Move MediaTrackConstratintSet into its own file.
