+++ WebRtcNet.Api/WebRtcNet.Api.csproj
@@ -1,15 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net10.0;net48</TargetFrameworks>
-    <OutputType>Library</OutputType>
-    <Configurations>Debug;Release</Configurations>
-    <Platforms>AnyCPU;x64;x86;ARM64</Platforms>
-    <AssemblyTitle>WebRtcNet.Api</AssemblyTitle>
-    <AssemblyDescription>The public WebRtc interfaces and classes for the WebRtcNet library</AssemblyDescription>
-    <Guid>98433217-c99b-4b08-a19b-3f97a1f1e633</Guid>
-    <DocumentationFile>bin\$(Configuration)\WebRtcNet.Api.XML</DocumentationFile>
-    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
-	  <LangVersion>latest</LangVersion>
-    <RootNamespace>WebRtcNet</RootNamespace>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFrameworks>net10.0;net48</TargetFrameworks>
+		<OutputType>Library</OutputType>
+		<Configurations>Debug;Release</Configurations>
+		<Platforms>AnyCPU;x64;x86;ARM64</Platforms>
+		<AssemblyTitle>WebRtcNet.Api</AssemblyTitle>
+		<AssemblyDescription>The public WebRtc interfaces and classes for the WebRtcNet library</AssemblyDescription>
+		<Guid>98433217-c99b-4b08-a19b-3f97a1f1e633</Guid>
+		<DocumentationFile>bin\$(Configuration)\WebRtcNet.Api.XML</DocumentationFile>
+		<AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+		<LangVersion>latest</LangVersion>
+		<RootNamespace>WebRtcNet</RootNamespace>
+	</PropertyGroup>
 </Project>
